Rename configuration to `KiaiFlashIntensity`
