+++ osu.Game.Rulesets.Taiko/Skinning/Argon/ArgonCirclePiece.cs
@@ -104,7 +104,7 @@ namespace osu.Game.Rulesets.Taiko.Skinning.Argon
         [BackgroundDependencyLoader]
         private void load(OsuConfigManager config)
         {
-            config.BindWith(OsuSetting.KiaiFlash, flashOpacity);
+            config.BindWith(OsuSetting.KiaiFlashIntensity, flashOpacity);
         }
 
         protected override void OnNewBeat(int beatIndex, TimingControlPoint timingPoint, EffectControlPoint effectPoint, ChannelAmplitudes amplitudes)
