+++ osu.Game.Rulesets.Taiko/Skinning/Default/CirclePiece.cs
@@ -150,7 +150,7 @@ namespace osu.Game.Rulesets.Taiko.Skinning.Default
         [BackgroundDependencyLoader]
         private void load(OsuConfigManager config)
         {
-            config.BindWith(OsuSetting.KiaiFlash, flashOpacity);
+            config.BindWith(OsuSetting.KiaiFlashIntensity, flashOpacity);
         }
 
         protected override void LoadComplete()
