+++ osu.Game/Configuration/OsuConfigManager.cs
@@ -161,7 +161,7 @@ namespace osu.Game.Configuration
             SetDefault(OsuSetting.ScalingPositionY, 0.5f, 0f, 1f);
 
             SetDefault(OsuSetting.UIScale, 1f, 0.8f, 1.6f, 0.01f);
-            SetDefault(OsuSetting.KiaiFlash, 0.3f, 0.10f, 1.0f, 0.01f);
+            SetDefault(OsuSetting.KiaiFlashIntensity, 0.3f, 0.10f, 1.0f, 0.01f);
 
             SetDefault(OsuSetting.UIHoldActivationDelay, 200.0, 0.0, 500.0, 50.0);
 
@@ -273,7 +273,7 @@ namespace osu.Game.Configuration
                         // TODO: implement lookup for framework platform key bindings
                     )
                 ),
-                new TrackedSetting<float>(OsuSetting.KiaiFlash, strength => new SettingDescription(
+                new TrackedSetting<float>(OsuSetting.KiaiFlashIntensity, strength => new SettingDescription(
                         rawValue: strength,
                         name: GraphicsSettingsStrings.KiaiFlash,
                         value: $"{strength * 100:N0}%"
@@ -366,8 +366,7 @@ namespace osu.Game.Configuration
         ScalingSizeX,
         ScalingSizeY,
         UIScale,
-        KiaiFlash,
-
+        KiaiFlashIntensity,
         IntroSequence,
         NotifyOnUsernameMentioned,
         NotifyOnPrivateMessage,
